habilitacion dde mejoras para manejo de graficos en el front

+++ src/models/operations/indicators/tabulation.models.js
@@ -21,7 +21,7 @@ const Tabulation = db.define('tabulation', {
     allowNull: true,
   }, 
   percent: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.FLOAT(8,2),
     allowNull: true,
   }, 
 });
